requirements.txt
Flask==3.0.0
pulumi==3.94.2
pulumi-aws==6.10.0
pulumi-gcp==7.2.0
python-dotenv==1.0.0
